Created Volunteer Positions Pages
Turns out I was on the right track of how to create it. I went back to when the Admin Info page was created and then scaffolded it there, Then I went through and ignored anything that wasn't important to create a .gitignore. All that was left after this were the things in the folder that was just created. Seeing that this didn't change anything outside of that, I copied the folder to my desktop, discarded all changes, pulled to the latest version, and then dragged the folder in.
IT WORKS!!!!!!!
Now I just have to implement the table into VolunteerCreate which is more difficult than I thought, but there are some resources online that I found that should help me do that.